price-tracker/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
apscheduler==3.10.4
playwright==1.49.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
